backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
bcrypt==4.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
deepface==0.0.95
Deprecated==1.2.18
dnspython==2.7.0
filelock==3.19.1
fire==0.7.1
Flask==3.0.0
Flask-Compress==1.14
flask-cors==6.0.1
Flask-Limiter==3.5.0
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==21.2.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.3
libclang==18.1.1
limits==5.5.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.5
numpy==2.1.3
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==5.29.5
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.6.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
retina-face==0.0.17
rich==13.9.4
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
tf_keras==2.19.0
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
